OBUDPST-2: Increase accuracy of rate calculations
To increase the accuracy of trial and sub-interval rate calculations,
switch from milliseconds to microseconds for delta time values.
Because this modifies the scale of the values exchanged between client
and server (and would cause inaccurate rates at one end or the other)
the protocol version is also incremented to prevent tests with the
previous version.