OBUDPST-5: Display option for bimodal maxima
When the '-i <count>' option is specified the software will keep track
of and display two maximum rates. The <count> parameter specifies how
many of the initial sub-intervals comprise the first mode, while all
the remaining sub-intervals will be considered part of the second. For
better clarity when associating a range of sub-interval samples with a
maximum value (particularly when the sub-interval period is greater
than one second), a sequence number has been added to the displayed
sub-interval text.