OBUDPST-10: Option to ignore out-of-order
When the '-R' option is specified the software will ignore
sequence errors due to out-of-order and duplicate datagrams.
When a "late" datagram is received, the loss count is now
also decremented to adjust for the loss count that was
previously incremented when it was not where it was expected.
The test for duplicates is done by checking any lower than
expected sequence numbers against a history buffer of recently
received values.