OBUDPST-21: Timer option for older systems
This option allows the software to be compiled for older
client devices that are unable to support the required clock
resolution. When DISABLE_INT_TIMER (udpst.h) is defined a
system interval timer is no longer utilized. Because this
results in high CPU utilization, it is not recommended for
standard server operation.