Multi-Flow and Random Sizes
Support for multiple connections (i.e., UDP flows)
between the client and one or more server instances
was added (OBUDPST-51). This includes server
instances on the same machine or physically
distributed on separate machines, VMs or containers.
To support multiple connections to individual
instances and allow for some instances to be
unreachable or unavailable, the new command-line
option '-C cnt[-max]' was added. See README for
details and examples.

Additional changes included reducing the starting
sending rate while adding support for randomized
packet sizes (OBUDPST-48), stopping attacks that
prevent graceful test shutdown by manipulating STOP
bits in load PDUs (OBUDPST-49), adding an optional
'-1' (dash one) flag to the command-line that will
cause the server to cleanly exit after a single
test (OBUDPST-50), and adding support for GSO and
RecvMMsg()+Truncation optimizations (OBUDPST-52).
Testing showed that RecvMMsg()+Truncation performed
better than GRO, so it was used instead.

A byproduct of the above changes was that new
fields were required in the setup and load PDUs.
As a result, both the current and minimum protocol
version needed to be bumped to 10.