Optional Rate Limiting
To better support server scale testing, an optional
rate limiting mode can be enabled and compiled in
(see README for details). This mode takes advantage
of the existing bandwidth management option
'-B mbps' by having the server limit the maximum
sending rate the rate adjustment algorithm can use.
As such, this mode only needs to be enabled (and
the software recompiled) on the server to take
effect. Otherwise, the bandwidth management option
continues to only be used for test admission control.