OBUDPST-56: Interface Stats
This change addresses the need to always handle interface
byte counters as 64-bit values, regardless of the machine
architecture. Some compiler warnings were cleaned up and
the server will now display a message, when verbose is
enabled, if an invalid setup request is received. Max
requestable bandwidth on client (-B mbps) increased from
10000 to 32767. The 64-bit ntohll/htonll macros were
modified to accommodate big-endian architectures and a
column for lost status feedback was added to export.